sth/querier-tweaks merge
